requirements.txt
tensorflow==2.18.0
Flask==3.0.3
gunicorn==22.0.0
Pillow==10.4.0
numpy==1.26.4